atf_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
python-multipart==0.0.12
